Fix useEffect
Fixing the following warning:
eslint: warning react-hooks/exhaustive-deps - React Hook useEffect has
missing dependencies: 'betaPassword' and 'setAppUser'.
Either include them or remove the dependency array.
Fixed the warning by moving the functions inside of the useEffect since
they're not being used outside of it and passing in the betaPassword
in the dependency array. I figure that it's ok if the beta causes
the user stuff to run again since all the beta stuff will be going
away on launch.