Add AdminRoute
Added an AdminRoute that checks if the App's admin property is true
to render the route's component. Otherwise it will redirect. Doing this
to make the code more DRY by hoisting the admin checks from the admin
components up to the App. This simplifies those components by making
it so that we don't have to use the context 'n stuff.