Update User to Recipient
We're changing the "users" to "recipients" because it makes more sense.
It's clearer for the client/end-user that these are the people they're
sending gifts to as opposed to like, users that can log in or whatever.
Same with "event" vs "gift." Gift fits better with the product and
Event was a bit vague on one-time vs. recurring.