Add Gift Date ToolTip
Added a ToolTip for the Gift Date for One Time Gifts that describes what
that field means. Added flexbox styling to the span to get the icon to
line up nicely with the Gift Date text. Otherwise the icon looks like
it's way too low compared to the Gift Date text.