Remove console.log
Removing the console.log calls in the files I'm working in to clear
up the console a bit. As for the server requests, keeping those logs
but changing them to the `log` function from App.tsx so that they
don't show up in the console on production, since that's usually
not a good idea.