Break Out Event Dashboard useEffect
Separating the useEffect in eventDashboard.tsx into each fetchRequest
so that we can properly populate the dependency arrays and remove the
warnings from the console. Also removing the Loading message since it's
just a line of text that pushes everything down and doesn't really
do a very good job indicating that anything is really loading. I think
we'd want to have each dashboard have its own spinning icon if it's
waiting for data.