Fix Red Console Warnings
Fixing three angry red console warnings on the event/Edit Gift page:
<div> cannot appear as a descendant of <p>.
Each child in a list should have a unique "key" prop.
<ol> cannot appear as a descendant of <p>.
By changing the paragraph tags to div's, we can have block elements
inside without it getting angry about trying to put them in an inline
paragraph tag. But in another case, we need to change the div's to
spans because they're inline elements that CAN go inside of paragraphs.
Finally, using the department string as a key since it should be
unique. And if it's not...well, that's their own fault.