Remove eventNew console.log
Removing the extra console.log's from the eventNew.tsx since we don't
need those in the repository. If a developer is working on the code,
they can add them for debugging, but we shouldn't be committing them.
If we do keep them, we should make sure they don't show up in prod,
so changing them to the `log` function from App.tsx that checks if a
debugging flag is set first.