Change show to showModal
Changing the useState from show/setShow to showModal/setShowModal
so it's more clear exactly what it's showing. There's also showItemList,
so it helps to be specific in this case. Also changing "Edit Gift" to
"Save Gift" since that's a bit more intuitive that you need to press
the button to get your changes to save.