Remove Promise.all
Changing a couple .map functions to .forEach because the functions inside
aren't returning anything (you get an array of "undefined"). I think
the idea of trying to use the Promise.all was to wait for the .map's
to finish, but since they aren't doing anything asynchronous, all of the
code has completed by the time the Promise.all runs so it's not doing
anything. This removes a couple warnings that say
"Array.prototype.map() expects a return value from arrow function."