Remove "onetime" Filter
There used to be a check in the "Current Gifts" list that would make
it so that if the name of the event was "onetime" then it wouldn't
show up. I confirmed with Krista and Victoria that one time gifts
are supposed to show up in the list, so got rid of the check.