Break Out eventNew useEffect
Breaking out each of the useEffect functions into their own useEffect
hooks so that they can each have their own dependency lists. Removes
warnings that say "React Hook useEffect has missing dependencies:...
Either include them or remove the dependency array."
The functions that are only called by the hooks are moved into them so
that the linter can determine what needs to be in the dependency array.
Also moved the useState block to the top with the rest of them. Keepin'
'em together.