Fix Last eventNew useEffect
Fixing the last eventNew useEffect hook by moving the handleInitialEventSetup
function into it. It was being called in another place in the code,
but it was just immediately returning the event, so I just copied over
that small part so that we could keep the function inside the hook and
know what our real dependency list is from the linter. Also moved
the last useState hooks up to the top so they're all in the same spot.
It seems like the massive amount of useState's would indicate we need
to break up some stuff into their own components...