Fix Logout Route
All of the routes are relative hash routes, so an absolute /logout
path doesn't work. Changing it to /#/logout so that it works correctly.
Also removing unused variables and moving the logout function in the
logout.tsx component into the useEffect so that we don't have to pass
it in the dependency array (fixes the linter warning).