Clean Up userDashboard
Ran JSPrettier on userDashboard and removed all of the unused variables,
functions, and imports. Also updated the useEffect dependency list so
that the linter isn't angry and moved the settingUsers function into it
so that we don't have to include it in the dependency list (it's not
being used anywhere else, so we're good to move it into the hook).
Also removed commented out code. We can always find it again in the
repo history.