Handle Local and ISO Dates
For recipient user Birthday and Start Dates, making it so that we support
both ISO UTC strings as well as local string types. This is so that we
make sure that we're not randomly changing timezomes between front-end,
backend, and data models. The rule of thumb should be that dates are
ISO strings in the UTC timezone except when we display them. This
commit tries to get closer to that ideal, but the data in the database
isn't quite there yet, so we need to handle edge cases...