Style Recipient Detail Rows
Adding the button role to the userListContainer detail rows so that
the cursor turns to a pointer and it's more apparent to the user that
they can click on the row like a button to edit the recipient details.
Also adding a class with a hover style for the row so it's more
responsive looking when you're using a mouse.