Style Selected GroupedItemCard
In the EventNew component (the event page), made it so that the selected
GroupedItemCard is based off of the activeItemDetails instead of the
activeItem so that it will apply the classes immediately instead of
waiting until the user presses the "Select Gift" button. This lets them
see more easily which package they've currently selected (i.e., which
one is showing details at the moment). Added a style to make the
non-selected ones a bit transparent. Also added the border to what looks
like was probably intended. And removed some console.logs. We don't want
to commit those.