Change Shipping Wording
Changed the "Shipping: TBD per person" on the eventNew.tsx to the
wording "Shipping to Recipient's Home Address" and moved the "cost"
portion over to be aligned with where the other cost is, so that
they'll match up later when we calculate it. Changed the wording because
that's where all gifts will be shipping for now. Later they might be
able to change it to different locations like the office.