Run JSPrettier on userAddRecipientContainer.tsx
Ran JSPrettier on userAddRecipientContainer.tsx. It fixed some whitespace
in the JSX and did some small stuff like replace single quotes with
double. Also changed the wording for "Edit Recipient" to "Save Changes"
so that it's more apparent that they need to click the button for their
changes to be saved. Eventually it would be nice if the changes were
saved without them having to press a button because that's what they're
used to in their phones and stuff. But then we'd need to implement a
debounce, so we'll wait on that.