Add GraphQL
Added in the GraphQL Amplify API. Because one of the paths was absolute,
we ended up changing it to a relative path and added a note in the
README about how to change it. Hopefully they fix it soon and it will
just work like this and we can get rid of the README note. That'd be nice.
Right now, it's just using the default Blog Template for us to make sure
it's all working 'n stuff.