Change to Todo
Changed the schema from the blog template to the todo template so that
we can finish the tutorial. We also needed to add the local resolver
directory to the .gitignore because we didn't want to accidentially add
those. If we ever want to overwrite the auto AWS resolvers, we'll want
to remove that .gitignore line, but for now this will keep us from
making mistakes. Also added the front-end code. It's real spiffy!