Create GraphQL Schema
Created the graphql schema from our design document for the database
architecture. Was unable to run amplify mock api because it said
our cognito didn't have user pools set up yet, so going to have to
figure that out before we can verify that this is a valid schema or not.