Add Auth User Pool
Added the Cognito user pool group Admin so that the schema will work.
Also changed the public access to private in the schema so that
the users still have to be logged in to see the data, otherwise we also
need to turn on iam for public access and I figure that we can figure
that out later. For now, keeping things more simple. Did an amplify
pull to get the updated auth stuff and committing that, too.