Add userPoolGroups to Backend Types
The amplify-dependent-resources-ref.d.ts was updated with a userPoolGroups
section that wasn't there before. Not sure when it got added, but it
probably should have gone in with the last commit. Oops. Also updated
my local node from version 14 to 17 and so updating the package-lock.json
as well. Husky wasn't working, and that was very bothersome...