Add back Todo
Adding back the Todo to the schema so that we can keep using it for
testing for now and also because I need to push the auth to the cloud
to be able to test the new schema locally with amplify mock api.
It looks like it also added the User Pool to the cli-inputs.json?
Why does it keep updating these files after I try to push...? I wish
they'd be all set up before I try and go to commit...