Update GraphQL
Finally got the amplify mock api to work on my local machine. Played
around with it a bit and it seems to work! Updating the autogenerated
queries, mutations, and subscriptions. Also Changing the default auth
to be the Cognito User Pools and giving it the ID of the user pool
that's attached to the ZipZap-dev Cognito.