Add .prettierignore
Adding a .prettierignore file that's a duplicate of the .eslintignore.
The .eslintignore is the recommended one for the react amplify project,
and we don't want prettier to change the formatting of the files
in src/model because those are autogenerated and it's a pain to have
them be changing every time.