Style Sidebar
Styling the sidebar to match the designs. More or less. It's lookin'
pretty nice, thooough. Also added a default catch-all route so that
we can see the functionality of the links better for when they're the
actively selected link. Opted to use the CSS Hover instead of doing
it in the JavaScript since that would need to add a hover state and
use the MouseOn/MouseOff property event things, which looked like a pain.
We might want to look into the magical &.hover support for inline CSS.