Add Mobile Top Bar
Added the top menu bar for mobile so that the user can click the menu
button to pop out the menu from the side. The top mobile bar has the
Zip Zap logo and a menu button. Added the media queries to the index.css
because we don't have support for those as inline styles right now.