Close Menu on Window Resize
Closing the menu whenever the window is resized because when the screen
size changes it does a wonky jump to go back to the other styles. Also
moving the animation to when it actually changes to Active because
otherwise during the screen size change it will animate to the other
position and look a bit silly.