Add Sidebar Branding
Added the place holder for the user's profile picture in the sidebar.
We'll need to update it later to actually support using the user's
uploaded profile picture instead of just the default_photo.png.
Also added the logo and made it so that it doesn't change size when you
click on the Gift Dashboard link.