Designing login
Things I have tried: if you add the component before the authenticator is stays there even after logging in. Working on finding a state variable that can be used for this.
You can customize the css variables, but I have not found the right variables to give for the primary