front end changes
started to work on importing the event dashboard but I can't find the new styled page.

Adding front end to profile:
using the context obj to pull the name phone and email.
still need to make a way to add address, and company.
Also need to add the stripe api to see the credit cards on file