Prettier Updated Files
Ran prettier on all of the files that have been updated on the master
branch. This is so that we have consistent code that looks pretty and
makes it easier to read and debug. Also removed extra code the linter
was warning about and some console.logs. Also bringing back husky.