Clean Up Todos
Cleaning up the react-query and todo stuff so that we're exposing as
little as possible of the react-query in the Todo component so that
we keep it as clean as possible. It would be nice to abstract it as
much as possible so that if anything changes around getting or updating
the data we don't have to change as many things.