Add GitHub Actions Workflow
Adding in a GitHub Actions Workflow that connects to amplify and then
runs the npm build and test to make sure that all of our code is good.
By automating this, we'll save a lot of time during code reviews since
we were wasting a lot of time doing that before. This will protect
us from silly things sneaking into the code. We might need to change
the amplify_env to prod later? Also adding a line in the .gitignore
to ignore the .secrets file for local secrets we can use to run
the github actions locally with the act command.