Fix GiftDashboard Linting
Fixing the npm run build warnings that come up for the GiftDashboard
because in the CI environment they cause the build to totally fail.
This is a good thing, because we don't want silly linting problems
to sneak into the codebase. This commit removed unused variables and
imports and stuff and makes the dependency array for useEffect super
specific to what it actually needs.