Remove Unused Stuff
Fixing all of the linting errors for unused variables and imports.
They clutter up the code and make it hard to be sure what's actually
being used when you have to go back and work on it. It makes maintaining
the code easier. Also moving the function used in the ProfilePage to
inside the useEffect function since that's the only place it's being
used and it was causing nasty spiral that the linter was warning about.