Include Dev Packages in Workflow
We need to include the devDependencies in the GitHub Actions workflow
because it includes prettier which it uses in the npm run build to
do eslinting stuff. I guess it makes sense, since build steps are
dev stuff....I wonder why it worked with act on my local but not on
GitHub itself. Huh....