Style Recipients List a Little
Doing a bit of styling for the recipients list so that it matches the
design a little bit more. It's still not very close, but it's better
than before. I want to talk with Krista before going much further on
it because I think we need to nail down if we're using Amplify
components or Bootstrap or whatever.