Fix Login Page
Making it so that the Login Page is mobile first and looks good on small
screens. Removing the App.css because we shouldn't be using any of that
CSS anymore, since we're using the MUI theme stuff (there's a couple
places that need className removed). Also preloading the
Poppin font so that there isn't that flash.