Add Gift Category
Adding the gift category as a string to the Gifts so that we can organize
them on the front-end into categories. Doing it as a string for now
so that we can iterate faster. We probably want to make it its own
queryable type later, but for now this will work.