Add UploadGift Component
Adding the UploadGift component which will handle creating new gifts.
This is an admin-only feature, but for now we're just worrying about
getting it out so that we can more easily create gifts since trying
to do it from the Amplify UI is real tricky...We're probably about
halfway done. We still need to add support for Gift Items and
uploading Images.