Remove Conflict Resolution
Turning off the Conflict Resolution in the GraphQL API because we aren't
using the DataStore stuff, so since we're using the pure API we want
to make sure that things are acting the way we expect, like deleted
things not coming back. We want them to actually delete themselves and
not hang around. We can look into DataStore later.