Make Gifts and Items Many to Many
Made Gifts and Items in a Many to Many relationship because I realized
that when creating Gift objects, we'll probably want to reuse Item
objects. For example, if we have Large, Medium, and Small Birthday
Boxes, all of them probably have the same Item 'Party Hat.' Before,
an Item could only belong to one Gift Box at a time.