Update Tables
Updating the example data in the TemplateTable so that it's closer
to the Figma data because we're going to be showing this to a potential
investor soon and we want it to look as professional as possible.
Also making it so that the RecipientList table's rows are hoverable
and it changes the cursor to a pointer so that users can tell that they
can be clicked on.