Format Address2
Changing up the formatting for address2 in the Recipient Profile's
GeneralInfo display so that if the address2 isn't filled in, it doesn't
have "null" and an extra comma. This is because many residential addresses
don't have an address2 field filled out and we want it to look as
professional as possible for potential investors.