Fix Recipient Loading
The recipient wasn't loading properly on the RecipientProfile.js page
because the useQuery function expects a function that returns a Promise
as its second parameter, not a raw Promise. So wrapped the GraphQL call
into a function and got rid of the map over an array of data since we're
only getting one recipient now.