Resize Calendar Events
Resizing the Calendar Event icons so that they're big on a big screen
and the default font size everywhere else. They still don't look perfect
on smaller mobile, but it does give that extra little bit of character
that will do for now. We need to nail down the design for these first,
but I did it good enough for now just in case I can't work on it for a
while.