Fix Department Error
Fixing an error that was breaking the app when the Department was
properly filled out for a Recipient. This is because the Department
is an object and React freaks out if you try to render an object. Making
sure to just grab the name instead. Also changed the Christmas dummy
picture because it was too big and taking too long to load.