Update Text
We're updating the text to match a bit better for presentation purposes.
"Default Gifts" don't make much sense, so quick swapping them out with
"Upcoming Event" things. We'll want to refactor the component to be better
named and stuff when we go to plug it in with data pulled from the database.
Also fixing the spelling of "Custom."