Replace Confetti
Removing the react-confetti-explosion package with react-rewards. Doing
this because react-confetti-explosion had a build problem where it had
some Warnings, and because it would run the explosion on every render.
Forced rerendering by react-query refreshing the Recipient List in the
background when we added a new user was causing more than one explosion
and looked a bit silly. The new confetti package only does the confetti
when we call the function, so we make sure that it only runs once.