Add Search
Added Fuse.js, a fuzzy text searcher, and used it in the Recipients
List to be able to filter recipients. This should let users be able
to find the people they want to edit much more easily. It puts preference
on their names, but also allows searching for things like birthdays and
job titles. We'll run it through some user testing and see what they like
best in the end, though.