Add Choose A Gift Search
Added functionality to the Choose A Gift search box. Replaced it with
a Material UI TextField and used Fuse to support fuzzy searching by
name or category. Also added styling for when the gift is selected, so
the user can more easily see which gift they selected, and also have it
automatically go to the next page after they select so that they don't
have to scroll all the way down to the Next button after they've selected.
If they got it wrong, they can always press the back button to go back.