Hoist Form State
Putting the entire Send a Gift form state into the top component so that
it can keep track of all of the state for each of the steps. This makes
it easy to go between steps and have the state be saved, and more importantly
it will make it easy to reload the gift if we need to pull it in later
to edit.