Save Recipient State
Making it so that when recipients are selected in Send A Gift, they're
saved to the formState. Also clearing out the orderDateType if they
change to ONE_TIME gifts. This is in addition to supporting showing
the correct date with the user if they select Birthday or Anniversary
for the Date Type.