Use Real Data in Check Out
Hooking in some real data to the Send A Gift's Check Out page. Needed
to update the GiftModal's setSelectedGift to return the whole gift
so that we'd be able to grab the image and price from it. The setInput
is getting a little hairy...still need to do all the Credit Card stuff...