Remove Header Import
Removed the Header import because it's not being used in SendAGift
anymore because we wanted the extra space on the page to help show the
Next button to the user. If we keep the import without using it, it
throws a Warning which fails the tests. Also added an apostrophe to
"Recipient's Address"