Add Email Defaults
For the "Want Custom Packaging?" email, made it so that it has a auto-
filled subject and first couple lines to help prompt the user into what
they should say. We probably want to turn it into a form later so we
can send additional information like the Company and Gift IDs. But this
will work for now. Gives us the opportunity to reach out to the person
and do up-selling and stuff for now.