Update Amplify
Pushed the new function updates that change stripe to use modules and
then did an amplify pull. It appears to have deleted the settings for
securing the REST API Endpoint behind auth, but....we'll see, I guess...
Ended up reverting the changes and then reverting the revert because
the problem turned out to be mocking the function locally as a module
and not pushing it into AWS. Still need to test it, but committing this
code because I'm going to be switching to a Windows lappy.