Update Stripe Function
Updated the stripe function to properly use ES6 Modules in lamda. The bad
news is that it doesn't support doing local mock functions, but with the
auth that we're doing, we probably don't need to worry about it in the first
place...It looks like it also brought back the file that is supposed to
protect the endpoint behind auth. Yaaay....