Finish Stripe Secret Endpoint
Finished the stripe /secret endpoint. Made it so that the user passes along their
GraphQL auth token so that we can use that to access the GraphQL data from
the backend as that user so they can update themselves. That being said, the
id that we're using for the user is the cognito identity context of the call
to the AWS lambda function. This will make it so that the user will only be
able to get a client secret for themselves. This is an attempt to make Payment
related things secure and not give other people the opportunity to get data
that isn't theirs.