Add Stripe /payments Endpoint
Added a new /payments endpoint to the stripe API lambda function so that
we can get all the payment methods for a user and let them select a saved
one the next time they check out. Trying to make it as easy as possible
to get their money, hur hur hur...Moved all the code to get the StripeID
from GraphQL into an auth.js file so that it's easy to re-use the code
and keep things clean.