Only Return Card Information
Updated the stripe /payments endpoint to only return the couple of details
we need on the front-end to display the credit cards and allow the user to
re-use them. I think in the future we'll want to use the metadata for
saving a name to the card, but this is all we need for now.