Add Ability to Add New Card
Giving users the ability to add another card even if they already have some
saved so that they can give us their best payment information so we can
get money. Also showing the circular loading icon when the Payment stuff
is loading and not showing the cancel button until it's ready so that the
user has a better experience and can't accidentally press the button before
it's ready.