Update GraphQL Order Schema
Updating the Order schema for the GraphQL model so that it has all of the
fields we need to populate the SendAGift state. This will make it more
easy to hydrate the state from the database if the user ever wants to update
any of the gift details, such as adding or removing recipients.