Create Gift Order
Added in the code to take the state from the SendAGift and create a real
Order. WooHoo! Did some minimal error catching to let the user know to try
again, and we still need to put in a TON of validation so users don't do
dumb things. But in the meantime, we've got the happy path data saving.
Yaaaaay! Also had to make sure to set the totalPrice on the formState so
it would be saved properly to the database. Accidentally forgot that. Oops.