Remove Unused Config
Doing an amplify pull removes the amplify\backend\api\APIGatewayAuthStack.json
file, and I verified that it's still removed when we run an amplify update
on the api and select to only allow Authenticated users. I believe that this
must have happened because we made that update after we had allowed the
api to be public, so now that it's behind the same cognito stuff as
everywhere else, it probably doesn't need that separate auth stack.