Display Gift Orders on Dashboard
Updating the Accordions on the Dashboard to show all of the Gift Orders.
Ordering them by date in Ascending order and moving ones that have already
passed to the bottom so that they're out of the way. Also only showing the
number of recipients and not their names since I don't think we have enough
room for it. We'd also need to grab their names....Also getting rid of the
"This Month" designation because it makes it hard to edit things in the
far future. We'll need to revisit the design to fix that...