Add Edit Button
Adding an "Edit Gift" button to the gift orders on the Dashboard so that
users can click on it to go back and edit their gifts. This will allow them
to add or remove users or change the name or payment method or whatever.
Had to use the text-decoration "none" on the react-router Link element so
that it wouldn't look like a regular html link with the underline.