Hydrate Send A Gift
If the user goes to a gifts/:id route, making it so that the formState is
hydrated with the data for that gift order from the API. Making sure to
only populate the formState with the order once when it comes back by
checking for the id on the formState, since a new form should never have
one of those. Also only running the query (enabled) if the params.id is set.
We also jump straight to the second step in the process if they are editing
and existing gift because they probably want to change the name or note or
recipients and not the gift itself. They can click the "Back" button if
they really want to change it.