Load Recipients
Turns out that if the user gets to the Choose Recipients page fast enough,
then the app will crash when it tries to loop through the recipients that
aren't there. This is much more likely to happen if they click on "Edit Gift"
and quickly click "Next" to edit them before the list has loaded. Oops.