Support Update Gift Order
Updated the addOrder function in the orders hook to do an update instead
of a create if it has an id. Also added an updatedBy field to the Order
so that we can see exactly who was the last person to update it. I think
that will come in handy. Last but not least, updated the last button in
the Send A Gift to say "Update Gift" if it's an update instead of a Create.
Also it apparently brought back the APIGatewayAuthStack.json...???