Default Gifts to Array
On the GiftCatalog component, we were getting the Gifts from react-query
but it can return an undefined if it's not done loading, so we were trying
to call the concat function and it would fail and kill the entire app. Oops!
Defaulting the gifts to an array so that we can safely call concat on it
no matter what.