Update Checkout Text
Updating the "Total Price" on the Checkout page to be Estimated Price because
we aren't going to charge their card until after we reach out to them and
finalize the details, such as shipping prices and any branding customization.
Also updating the success modal to let the users know that a Gift Expert
will be in touch with them soon.