Add Email Template
Added in the email template code that we'll be using for all of our Zip Zap
emails for now. Got the New Gift Order for Admin email set up and verified
that it looks good by copying the console.log from the AWS Cloud Watch.
We still need to add the Modified email and add a Client email as well.