Get Recipients With Lambda IAM
Updated the schema.graphql to support private iam access for reading recipient
data so that we can query it from our lambda function. Updated the checkoutemail
function to be able to use these new permissions to query for the users.
Still need to break them out and get their names and dates for the email,
but this at least shows us how to use IAM to get the data from GraphQL.
The documentation says that "IAM-based @auth rules are scoped down to
only work with Amplify-generated IAM roles" so hopefully that means that
user IAM we have to use for S3 Image stuff doesn't bleed into this and all
the Recipient data is secure behind the Cognito roles...hrrrngh...