Conditionally Get User Data
Added the "enabled" option to react-query to only run the query to get the
user if the userID isn't empty. This prevents other stuff down the line
getting angry if it's missing (GraphQL was getting really angry at an empty
string for the ID when it wasn't ready yet). Also removed the Credit Card
Section because it was breaking the page from rendering.