Fix useUsers Hook
Updated the useUsers hook to use the useQuery function correctly. Only
passing it the getUser graphQL string causes it to throw a "Missing queryFn"
error in the console. Also letting it take a userID to get the user, but
it should only run if it has a truthy value so it's not necessary to pass in.