Clean Up Profile Page
Cleaning up the Profile Page components so that we're only using the data
structure for Users from our GraphQL Schema. The old shape was used before
we had settled on the schema, maybe? It was causing confusion and problems,
so standardizing it across the components. Also needed to update the EditProfile
component to have the TextFields be controlled components. This meant making
sure that they didn't ever get "undefined" as their value, so defaulting
the fields to empty strings.