Update editUser
Updating the editUser function in the useUsers hook to handle saving edited
user information. This includes the user's company and company address info,
so added some helper functions to create or update a Company and Address
record whether or not they exist. Also had to use "phoneNumber" instead of
phone. Oops!