Differentiate Order Update
Making it so that the email has a different title if it's an update to a
gift order as opposed to a new one being created. This should make it easier
for the gift admins to see what they need to do until we can automate more
things. Also updated the console errors we log to be actual error types.
And including the name of the gift order in the email so the admins can find
it easier.