Send Checkout Email
Finally actually sending the checkout email! Added an identity in AWS SES
for zipzap.gifts so it can send the emails. For testing in the sandbox, also
had to add my own email address to the Custom Resource Policies to send it
to, but the internet says that once we're out of the SES Sandbox, we only
need identity permissions for the Source/From address. Tested it and it works!!
Unfortunately, turns out that the email body doesn't render the styles
like a raw html page in most clients (Gmail won't do the SVG logo or fonts,
Veranda is the closest to Poppins, and none of the background colors are applying)
so we'll have to figure out that later...