Add Real Admin Email
Using connect+admin@zipzapgifts.com as the administrative email for the
checkoutemail function. Adding it to the permissions for sending emails so
that we can use it while we're still in the SES Sandbox. We set up a rule
to have that email forward to everyone who needs to see it, because setting
up our zipzap.gifts domain to receive emails is a giant pain....