Update package-lock.json
Updating the package-lock.json because npm ci was failing on the Amplify
build. So deleted the node_modules folder and the package-lock.json, then
ran npm install again to regenerate the package-lock.json. Also ran npm ci
afterwards with npm run start to see if it would work and build on my local.
Here's crossing fingers that it works on Amplify...