Regenerate main env
Deleted the main env and regenerated it to make sure that the stripe api
and function is working. It was giving back 403s before. In retrospect,
it might have been a Cognito Group problem with the specific user, but
we'll test that out later. I'm just happy this is working for a minute...