Only Disable Next Button on Invalid Page
Making it so that the "Next" button on the Send a Gift page is only disabled
and marked as invalid when they're on the "Gift Details" step because otherwise
the user can press the "Back" button and the "Next" button remains grayed
out on the "Choose a Gift" step and they can't move forward. Oops!