Create User on LoginPage
Updating the LoginPage to use react-query and check if the logged in user
is already in the GraphQL database. If they aren't (GraphQL has finished loading
but we didn't get anything back) but the user is logged in with Cognito,
then we run a mutation to create the user in GraphQL using the Cognito
attributes. We have to use the Cognito username (it's a UUID) as the GraphQL
ID because we use that in some backend functions for authorization and
authentication on accessing other resources in AWS.