Display "Finish Setting Up Account" Message
If the user doesn't have a Company Name, we want them to finish setting Up
their profile by putting in their company name and address information.
We do this by grabbing their info from the GraphQL and seeing if they have
the Company Name filled out yet. We use the "enabled" react-query prop to
only run the query if we've successfully gotten the Cognito username (which
is the ID of the user).