Fix Destructure on null
Turns out that the default value for a destructuring only applies if the
item is totally undefined. So if the key exists but the value is null, it
won't assign it the default and we'll get errors if we try to destructure
on a null. So splitting out the nested destructuring and doing an || to
go with an empty object if the variable is null or undefined.