Add Lucky Orange
Adding Lucky Orange to the head tag so that it loads their script on our
site and does user tracking. Their dashboard looks like it has some good
filtering so we can make sure to filter out the localhost and dev stuff,
so not going to worry about only loading the script on live for now. Yaaay...