Add Login Email Template
Updated the cognito auth emailVerificationMessage to be the email template
that was designed for giving the user their login code and looking nice
and Zip Zappy. Removing the old file because it's not used.
It looks like this email template is referencing some old images in an
eastern coast S3 bucket...we'll need to move those eventually...