Save Favorites
Looping through all of the submitted favorites from the Google Form submit
event and saving them as ProfileFavorites to the database. Also updating
any existing favorites that have the same value. This means people can update
their answers, but if the questions even change a little bit, we'll need
to go back and clean them up. We can worry about that later.