Show Recipient Favorites
Updated the Recipient Profile's GiftProfile to display the saved ProfileFavorite
items attached to the recipient or to show a message that can let a User
re-send the Google Form to the recipient. I also removed the Gift History
tab because that won't be available for quite a while...And got rid of some
extra sx props that were redundant because the props apply from the smallest
up, so you only need to include the next breakpoint up where it should change.
Also changed the "Suggested Gifts" to say "Coming Soon!"