Add recipients api
Added a recipients api with the endpoint /favorites/survey that will send
out emails to the Google Form survey so that recipients can submit their
favorites to Zip Zap Gifts. Right now it's grabbing the email addresses for
user ids passed in by a user from the front-end. Doing it this way so that
we make sure to only send out emails to recipients that the requesting user
actually has access to.